third_party/blink/web_tests/wpt_internal/script_tools/provide_context_mixed_registration.html

<!DOCTYPE html>
<html>
<title>Script Tools: mixing individual and batch tool registration</title>
<link rel="author" href="mailto:brwalder@microsoft.com">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
promise_test(async () => {
  async function echo(obj) {
    return obj.text;
  }

  async function reverse(obj) {
    return obj.text.split('').reverse().join('');
  }

  async function upper(obj) {
    return obj.text.toUpperCase();
  }

  const automationDelegate = window.automationDelegate;

  // First register tools via provideContext
  automationDelegate.provideContext({
    tools: [
      {
        execute: echo,
        name: "echo",
        description: "Echo tool"
      },
      {
        execute: reverse,
        name: "reverse",
        description: "Reverse tool"
      }
    ]
  });

  // Then register individual tool
  automationDelegate.registerTool({
    execute: upper,
    name: "upper",
    description: "Uppercase tool"
  });

  // Unregister one tool that was registered via provideContext
  automationDelegate.unregisterTool("echo");

  // Unregister the individually registered tool
  automationDelegate.unregisterTool("upper");

  // Clean up remaining tool
  automationDelegate.unregisterTool("reverse");
}, "can mix provideContext and individual tool registration/unregistration");

promise_test(async () => {
  async function oldTool() {
    return "old";
  }

  async function newTool1() {
    return "new1";
  }

  async function newTool2() {
    return "new2";
  }

  const automationDelegate = window.automationDelegate;

  // Register initial tools via provideContext
  automationDelegate.provideContext({
    tools: [
      {
        execute: oldTool,
        name: "tool_to_replace",
        description: "Tool that will be replaced"
      }
    ]
  });

  // Use provideContext again to replace the tools
  automationDelegate.provideContext({
    tools: [
      {
        execute: newTool1,
        name: "new_tool1",
        description: "New tool 1"
      },
      {
        execute: newTool2,
        name: "new_tool2",
        description: "New tool 2"
      }
    ]
  });

  // The old tool should be gone, only new tools should exist
  assert_throws_dom(
    'InvalidStateError',
    () => {
      automationDelegate.unregisterTool("tool_to_replace");
    },
    "old tool should not exist after provideContext replacement"
  );

  // Clean up new tools
  automationDelegate.unregisterTool("new_tool1");
  automationDelegate.unregisterTool("new_tool2");
}, "provideContext replaces previously registered tools");
</script>
